requirements.txt
fastapi
uvicorn[standard]
transformers
huggingface-hub
torchvision
pillow
python-multipart
